requirements.txt
torch
torchvision
transformers
pandas
numpy
matplotlib
Pillow
PyYAML
nltk
